etcdctl: remove old stuff
