+++ etcdctl/command/cluster_health.go
@@ -37,10 +37,6 @@ func handleClusterHealth(c *cli.Context) {
 	client := etcd.NewClient(endpoints)
 	client.SetTransport(tr)
 
-	if c.GlobalBool("debug") {
-		go dumpCURL(client)
-	}
-
 	if ok := client.SyncCluster(); !ok {
 		handleError(ExitBadConnection, errors.New("cannot sync with the cluster using endpoints "+strings.Join(endpoints, ", ")))
 	}
